requirements.txt
pillow
faiss-cpu
numpy
torch
transformers
streamlit
python-dotenv